Module Js

exception Not_implemented of string

val notImplemented : string -> string -> 'a

type 'a t = < .. > as 'a

module Fn : sig ... end

type +'a null = 'a option

type +'a undefined = 'a option

type +'a nullable = 'a option

val toOption : 'a null -> 'a option

val nullToOption : 'a null -> 'a option

val undefinedToOption : 'a null -> 'a option

val fromOpt : 'a option -> 'a undefined

val undefined : 'a option

The same as empty Js.Undefined will be compiled as undefined

val null : 'a option

The same as empty in Js.Null will be compiled as null

val empty : 'a option

type (+'a, +'e) promise

val typeof : 'a -> 'b

typeof x will be compiled as typeof x in JS Please consider functions in Types for a type safe way of reflection

operators

unsafe_lt a b will be compiled as a < b. It is marked as unsafe, since it is impossible to give a proper semantics for comparision which applies to any type

unsafe_le a b will be compiled as a <= b. See also unsafe_lt

unsafe_gt a b will be compiled as a > b. See also unsafe_lt

unsafe_ge a b will be compiled as a >= b. See also unsafe_lt

nested modules

module Null : sig ... end

module Undefined : sig ... end

module Nullable : sig ... end

module Null_undefined = Nullable

module Exn : sig ... end

module Array2 : sig ... end

module Array : sig ... end

JavaScript Array API

module Re : sig ... end

Provide bindings to Js regex expression

module String2 : sig ... end

JavaScript String API

module String : sig ... end

module Promise : sig ... end

module Date : sig ... end

module type Dictionary = sig ... end

module Dict : Dictionary

module Global : sig ... end

Contains functions available in the global scope (window in a browser context)

module Types : sig ... end

Provide utilities for manipulating JS types

module Json : sig ... end

module Math : sig ... end

JavaScript Math API

module Obj : sig ... end

Provide utilities for Js.t

module Typed_array : sig ... end

module TypedArray2 : sig ... end

Provide bindings for JS typed array

module Float : sig ... end

Provides functions for inspecting and manipulating floats

module Int : sig ... end

Provides functions for inspecting and manipulating ints

module Bigint : sig ... end

Provide utilities for bigint

module Option : sig ... end

Provide utilities for option

module Result : sig ... end

module List : sig ... end

module Vector : sig ... end

Provide utilities for Vector

module Console : sig ... end

val log : 'a -> unit

val log2 : 'a -> 'b -> unit

val log3 : 'a -> 'b -> 'c -> unit

val log4 : 'a -> 'b -> 'c -> 'd -> unit

val logMany : 'a array -> unit

module Set : sig ... end

Provides bindings for ES6 Set

module WeakSet : sig ... end

Provides bindings for ES6 WeakSet

module Map : sig ... end

Provides bindings for ES6 Map

module WeakMap : sig ... end

Provides bindings for ES6 WeakMap