Module Js.String2

JavaScript String API

type t = string

val make : int -> char -> string

val fromCharCode : int -> string

val fromCharCodeMany : 'a -> 'b -> 'c

val fromCodePoint : int -> string

val fromCodePointMany : 'a -> 'b

val length : string -> int

val get : string -> int -> string

val set : 'a -> 'b -> 'c -> 'd

set s n c sets the character at the given index number to the given character. If n is out of range, this function does nothing.

val charAt : string -> int -> string

val charCodeAt : string -> int -> float

val codePointAt : string -> int -> int option

val concat : string -> string -> string

val concatMany : string -> string array -> string

val endsWith : string -> string -> bool

val endsWithFrom : string -> string -> int -> bool

val includes : string -> string -> bool

val includesFrom : string -> string -> int -> bool

val indexOf : string -> string -> int

val indexOfFrom : string -> string -> int -> int

val lastIndexOf : string -> string -> int

val lastIndexOfFrom : string -> string -> int -> int

val localeCompare : 'a -> 'b -> 'c

val match_ : string -> Re.t -> string array option

val normalize : 'a -> 'b -> 'c

val normalizeByForm : 'a -> 'b -> 'c

val repeat : string -> int -> string

val replace : 'a -> 'b -> 'c -> 'd

val replaceByRe : string -> Re.t -> string -> string

val unsafeReplaceBy0 : 'a -> 'b -> 'c

val unsafeReplaceBy1 : 'a -> 'b -> 'c

val unsafeReplaceBy2 : 'a -> 'b -> 'c

val unsafeReplaceBy3 : 'a -> 'b -> 'c

val search : 'a -> 'b -> 'c

val slice : string -> from:int -> to_:int -> string

val sliceToEnd : string -> from:int -> string

val split : 'a -> 'b -> 'c

val splitAtMost : 'a -> 'b -> limit:'c -> 'd

val splitByReAtMost : 'a -> 'b -> 'c

val splitByRe : string -> Re.t -> string option array

val startsWith : string -> string -> bool

val startsWithFrom : 'a -> 'b -> 'c -> 'd

val substr : string -> from:int -> string

val substrAtMost : string -> from:int -> length:int -> string

val substring : string -> from:int -> to_:int -> string

val substringToEnd : string -> from:int -> string

val toLowerCase : string -> string

val toLocaleLowerCase : 'a -> 'b -> 'c

val toUpperCase : string -> string

val toLocaleUpperCase : 'a -> 'b -> 'c

val trim : string -> string

val anchor : 'a -> 'b -> 'c

val link : 'a -> 'b -> 'c

val castToArrayLike : 'a -> 'b -> 'c