server-reason-react

This site is the documentation page and API references for server-reason-react, server-reason-react is a Native implementation of React's server-side rendering (SSR) and React Server Components (RSC) architecture for Reason.

Designed to be used with reason-react and Melange. Together it enables developers to write efficient React components using a single language, while building for both native executables and JavaScript.

Guides

Get started

What does universal code mean?

How to organise universal code

Exclude client code from the native build

Externals and melange attributes

Core API

Those are the core libraries of server-reason-react.

React The React library

ReactDOM The ReactDOM library

Next

Get started