Introduction

server-reason-react is an implementation of react-dom/server and some of React's internals in OCaml. Its purpose is to render HTML markup from the server for a Reason React application natively.

module Greetings = {
  [@react.component]
  let make = (~name) => {
    <div>
      <h1>
        {React.string("Hello " ++ name)}
      </h1>
    </div>
  }
};

let html = ReactDOM.renderToStaticMarkup(<Greetings name="visitor" />)
// <div><h1>Hello visitor</h1></div>

React and ReactDOM

server-reason-react provides React and ReactDOM modules with the same interface as reason-react, including the JSX transformation via server-reason-react.ppx.

React server-reason-react.react

ReactDOM server-reason-react.reactDom

server-reason-react.ppx needs to be a different ppx than reason-react-ppx to transform domProps into an interal representation of Props where the type is known ahead of time in order to render them on the server.

Add server-reason-react.ppx to your pps list in your dune files

(preprocess (pps server-reason-react.ppx)) 

Make sure your code is universal

One big challenge of sharing code between client and server is that the server and the client platforms have different APIs available. You can't use browser's APIs on the server, such as document.querySelectorAll and you can't use server related APIs on the client such as any filsystem operation.

In this aspect server-reason-react SSR is not much different than Node.js. Node.js doesn't provide a window/document/etc and leaves the user to manually check for them on each usage. In our case, we don't provide any implementation for the browser APIs but we want the code to compile.

There're a few utilities that will be handy to make your frontend code work in native.

server-reason-react comes with a few modules that are compatible with both server and client, to make it easier to write universal code and don't worry about the platform.

Belt is an implementation of Belt that would work on both server and client. server-reason-react.belt

Js is an half-implementation of the Js module from melange.js, and many parts aren't implemented and some other parts aren't possible to implement on the server (Unstable, can raise "NOT IMPLEMENTED"). server-reason-react.js

Webapi is a stripped down version of melange-webapi that will crash at runtime if you call those APIs on the server. server-reason-react.webapi

Promise is a vendored version of aantron/promise that will work on both server and client. server-reason-react.promise

Exclude client code from the native build

Even when Belt, Js, and other modules might be universal, sometimes you want to discard some code from native, and only execute it on the client. For example, if you're using Webapi to query the DOM, you only want to run that on the client.

Thanks to browser_ppx, we can easily discard what's meant to be running on the client and avoid the execution on the server:

let%browser_only countDomNodes = (id) => {
  let elements = Webapi.Element.querySelector("#" ++ id);
  let arr_elements = Webapi.Element.toArray(elements);
  Array.length(arr_elements);
};

Add browser_ppx to your pps in your dune files. Don't forget to add it on both "server" and "client" dune files, since you want to discard the content and add browser_only -js only to the client.

On client's dune:

(preprocess (pps browser_ppx -js)) 

On server's dune:

(preprocess (pps browser_ppx)) 

The method used by browser_only to discards the function is transforming the body of your function into a raising exception. Sometimes it can be useful to wrap it in a try/catch block to prevent the exception from crashing, or to provide a default value or do something else.

For example:

let%browser_only countDomNodes = (id) => {
  let elements = Webapi.Element.querySelector("#" ++ id);
  let arr_elements = Webapi.Element.toArray(elements);
  Array.length(arr_elements);
}

let main = id =>
  try(countDomNodes(id)) {
  | _ => 0
  };

Externals and melange.ppx attributes

Since melange.ppx is not compatible with native, we provide server-reason-react.melange.ppx instead. It allows to have the same features as melange.ppx, so all externals, mel attributes and other features are supported.

Supports all mel. attributes

mel.* attributes are stripped out of the native build, and transformed into raising functions to raise at server runtime.

Enables pipe_first ->

it's a syntax sugar for pipe_last ||>, but it's not supported by default in native. server-reason-react.melange.ppx enables it and works perfectly fine in native.

Supports RegExp %re "/regex/"

Transforms %re ... into Js.Re.t from server-reason-react.js and works perfectly fine in native (There might be some features missing in the implementation, such as backtracking or grouping.)

Debugger %debugger

It removes the debugger in native. It's a noop on the server context, and it's pretty uncommon to use it on shared code.

Supports Object access ##

Unstable. It's not supported by default in native, since ## operates on JavaScript Objects, and the interface of that isn't polished.

(preprocess (pps server-reason-react.melange.ppx))